rename: CookieUtils -> CookieProvider and move package
