+++ src/main/java/com/example/backend/controller/UserController.java
@@ -12,7 +12,7 @@ import com.example.backend.global.exception.ForbiddenExceptionType;
 import com.example.backend.global.security.AuthToken;
 import com.example.backend.global.security.CustomUserDetails;
 import com.example.backend.global.security.AuthTokenProvider;
-import com.example.backend.global.utils.CookieUtils;
+import com.example.backend.global.security.CookieProvider;
 import com.example.backend.global.utils.ResponseUtils;
 import com.example.backend.service.user.SnsUserService;
 import com.example.backend.service.user.UserService;
@@ -39,8 +39,7 @@ public class UserController {
     private final UserService userService;
     private final SnsUserService snsUserService;
     private final AuthTokenProvider authTokenProvider;
-//    private final ResponseUtils responseUtils;
-    private final CookieUtils cookieUtils;
+    private final CookieProvider cookieProvider;
 
     @Operation(summary = "회원가입",
             description = "unique field 중복 시 errorCode -101(이메일), -102(닉네임), -103(이메일,닉네임)이 반환됩니다.")
@@ -71,7 +70,7 @@ public class UserController {
         //회원가입 성공시 SNS 유저는 로그인에 성공한다
         AuthToken AT = authTokenProvider.issueAccessToken(createdUser);
         AuthToken RT = authTokenProvider.issueRefreshToken(createdUser);
-        ResponseCookie RTcookie = cookieUtils.createRefreshTokenCookie(RT.getToken());
+        ResponseCookie RTcookie = cookieProvider.createRefreshTokenCookie(RT.getToken());
         return ResponseUtils.getLoginSuccessResponse(createdUser.getId(), AT, RTcookie, userTypeStr + " 로그인에 성공했습니다.");
     }
